Move GET_CORPUS_VERSIONS query to queries.ts per project conventions
